implement authentication for all routes and change icon

+++ app/server/routes/notes.js
@@ -1,28 +1,29 @@
 const router = require('express').Router(),
+  { isAuthenticated } = require('../authentication/openid'),
   { addNote, updateNote, deleteNote, listNotes } = require('../datastore')
 
-router.post('/add', async (req, res) => {
+router.post('/add', isAuthenticated, async (req, res) => {
   if (req.user && req.body.content)
     res.sendStatus(await addNote(req.user, req.body.content))
   else
     res.sendStatus(400)
 })
 
-router.post('/update', async (req, res) => {
-  if (req.body.id && req.body.content)
+router.post('/update', isAuthenticated, async (req, res) => {
+  if (req.user && req.body.id && req.body.content)
     res.sendStatus(await updateNote(req.body.id, req.body.content))
   else
     res.sendStatus(400)
 })
 
-router.get('/delete/:id', async (req, res) => {
-  if (req.params.id)
+router.get('/delete/:id', isAuthenticated, async (req, res) => {
+  if (req.user && req.params.id)
     res.sendStatus(await deleteNote(req.params.id))
   else
     res.sendStatus(400)
 })
 
-router.get('/list/:user', async (req, res) => {
+router.get('/list/:user', isAuthenticated, async (req, res) => {
   if (req.user)
     res.send(await listNotes(req.user))
   else
